+++ src/main/kotlin/com/john/bookmanagementsystem/presistance/BookRepository.kt
@@ -3,4 +3,4 @@ package com.john.bookmanagementsystem.presistance
 import com.john.bookmanagementsystem.model.Book
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface BookRepository : JpaRepository<Book, Int>
+interface BookRepository : JpaRepository<Book, Long>
